requirements/base.txt
# Base Dependencies
Django==5.0.6

# Environment and Configuration
django-environ==0.11.2

# Database and ORM
psycopg2-binary==2.9.9

# REST Framework
djangorestframework==3.14.0
drf-spectacular==0.27.2

# Celery and Related Extensions
celery==5.4.0
django-celery-results==2.5.1
django-celery-beat==2.6.0

# Security
django-guardian==2.4.0
pyotp==2.9.0

# Audit
django-easy-audit==1.3.6

# Static Files and Media
whitenoise==6.5.0
pillow==10.3.0
django-storages==1.14.2
qrcode==7.4.2

# Utilities and Enhancements
django-filter==23.5
django-extensions==3.2.3
django-cors-headers==4.3.1

# Polymorphic Models
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10

# API Documentation
drf-spectacular==0.27.2

# Channels and ASGI
channels==4.1.0
daphne==4.1.2

# Deployment
gunicorn==22.0.0

# Development Tools
setuptools==70.0.0

weasyprint==62.3
django_weasyprint==2.3.0
django-browser-reload==1.13.0
django-tailwind==3.8.0

drf-extra-fields==3.7.0